requirements.txt
google-adk>=0.1.0
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0